some pages added the navigate

+++ src/pages/MyReviews/MyReviews.jsx
@@ -1,7 +1,7 @@
 import React, { useContext, useState } from "react";
 import { useEffect } from "react";
 import { UserContext } from "../../context/AuthContext";
-import toast, { Toaster } from 'react-hot-toast';
+import toast, { Toaster } from "react-hot-toast";
 import { Link, useLoaderData } from "react-router-dom";
 
 const MyReviews = () => {
@@ -9,94 +9,109 @@ const MyReviews = () => {
   const [res, setRes] = useState(true);
   const { user } = useContext(UserContext);
 
-  const serviceData = useLoaderData()
-  console.log(serviceData)
+  const serviceData = useLoaderData();
+  console.log(serviceData);
 
   useEffect(() => {
     fetch(`http://localhost:5000/add-review?email=${user?.email}`, {
       headers: {
-        authorization: `Bearer ${localStorage.getItem('token')}`
-      }
+        authorization: `Bearer ${localStorage.getItem("token")}`,
+      },
     })
       .then((res) => res.json())
       .then((data) => {
         // console.log(data)
-        
+
         setMyReview(data);
       });
   }, [res]);
 
-  console.log(myReview)
+  console.log(myReview);
 
   const hanldeDelete = (userReview) => {
     // e.preventDefault()
-    fetch(`http://localhost:5000/add-review/${userReview?._id}`, {
+    const modal = window.confirm(`you want to delete the review?`);
+    if(modal){
+      fetch(`http://localhost:5000/add-review/${userReview?._id}`, {
       method: "DELETE",
     }).then((data) => {
-      setRes(!res)
-      console.log(data)
-      toast.success('Successfully Deleted the Review!')
+      setRes(!res);
+      toast.success(`Successfully deleted the review`)
+      console.log(data);
     });
+    } else {
+      toast.error(`Didn't delete the review`)
+    }
   };
 
-  const hanldeEdit = (userReview) => {
+  const hanldeEdit = (userReview) => {};
+  let count = 1;
 
-  }
-
-  document.title = 'My Review'
+  document.title = "My Review";
 
   return (
     <div className="mx-auto container">
       <h1 className="text-center text-4xl font-bold my-4">My review</h1>
       <div>
         <div className="overflow-x-auto w-full">
-          <table className="table w-full">
+          <table className="table w-full my-8 border border-separate">
             {/* <!-- head --> */}
             <thead>
-              <tr>
-                <th></th>
-                <th>Name</th>
-                <th>Job</th>
-                <th>Favorite Color</th>
-                <th></th>
+              <tr className="text-center">
+                <th className="w-[1rem]">No.</th>
+                <th>Service Name</th>
+                <th>Your Review</th>
+
+                <th>Action</th>
               </tr>
             </thead>
             <tbody>
               {myReview.map((userReview) => {
-                const { displayName, email, photoURL, review, _id } =
-                  userReview;
+                const {
+                  displayName,
+                  email,
+                  photoURL,
+                  image,
+                  price,
+                  title,
+                  review,
+                  _id,
+                } = userReview;
 
                 return (
                   <tr key={_id}>
-                    
-                    <th></th>
+                    <th className="border">
+                      <span>{count++}</span>
+                    </th>
                     <td>
                       <div className="flex items-center space-x-3">
                         <div className="avatar">
                           <div className="mask mask-squircle w-12 h-12">
                             <img
-                              src={photoURL}
+                              src={image}
                               alt="Avatar Tailwind CSS Component"
                             />
                           </div>
                         </div>
                         <div>
-                          <div className="font-bold">{displayName}</div>
-                          <div className="text-sm opacity-50">{email}</div>
+                          <div className="font-bold">{title}</div>
+                          <div className="text-sm opacity-50 font-semibold">
+                            ${price}
+                          </div>
                         </div>
                       </div>
                     </td>
                     <td>
-                      <p>{review}</p>
+                      <p className="whitespace-normal">{review}</p>
                       <br />
                       <span className="badge badge-ghost badge-sm">
-                        Desktop Support Technician
+                        ${price}
                       </span>
                     </td>
-                    <td>Purple</td>
-                    <th>
-                    
-                      <Link to={`/my-reviews/edit/${_id}`}
+
+                    <th className="text-right">
+                      <Link
+                        to={`/my-reviews/edit/${_id}`}
                         onClick={() => hanldeEdit(userReview)}
                         className="btn btn-success text-gray-900 btn-sm"
                       >
@@ -108,29 +123,16 @@ const MyReviews = () => {
                       >
                         DELETE
                       </button>
-
-                      
                     </th>
                   </tr>
                 );
               })}
             </tbody>
-
-            {/* <!-- foot --> */}
-            <tfoot>
-              <tr>
-                <th></th>
-                <th>Name</th>
-                <th>Job</th>
-                <th>Favorite Color</th>
-                <th></th>
-              </tr>
-            </tfoot>
           </table>
         </div>
       </div>
 
-    <Toaster />
+      <Toaster />
     </div>
   );
 };
